+++ zoomimage-compose/src/commonMain/kotlin/com/github/panpf/zoomimage/compose/zoom/Zoomable.kt
@@ -174,7 +174,7 @@ internal class ZoomableNode(
         detectPowerfulTransformGestures(
             panZoomLock = true,
             canDrag = { horizontal: Boolean, direction: Int ->
-                val supportDrag = zoomable.checkSupportGestureType(GestureType.DRAG)
+                val supportDrag = zoomable.checkSupportGestureType(GestureType.ONE_FINGER_DRAG)
                 val canScroll = zoomable.canScroll(horizontal, direction)
                 val supportOneFingerScale =
                     zoomable.checkSupportGestureType(GestureType.ONE_FINGER_SCALE)
@@ -189,7 +189,7 @@ internal class ZoomableNode(
                         zoomable.checkSupportGestureType(GestureType.ONE_FINGER_SCALE)
                     val supportTwoFingerScale =
                         zoomable.checkSupportGestureType(GestureType.TWO_FINGER_SCALE)
-                    val supportDrag = zoomable.checkSupportGestureType(GestureType.DRAG)
+                    val supportDrag = zoomable.checkSupportGestureType(GestureType.ONE_FINGER_DRAG)
                     zoomable.logger.v {
                         "ZoomableState. zoomable. onGesture. " +
                                 "longPressExecuted=$longPressExecuted, " +
@@ -238,7 +238,7 @@ internal class ZoomableNode(
                     val supportTwoFingerScale =
                         zoomable.checkSupportGestureType(GestureType.TWO_FINGER_SCALE)
                     val supportDrag =
-                        zoomable.checkSupportGestureType(GestureType.DRAG)
+                        zoomable.checkSupportGestureType(GestureType.ONE_FINGER_DRAG)
                     zoomable.logger.v {
                         "ZoomableState. zoomable. onEnd. " +
                                 "centroid=${centroid.toShortString()}, " +