+++ zoomimage-compose/src/commonTest/kotlin/com/github/panpf/zoomimage/compose/common/test/util/KeyHandlerTest.kt
@@ -1,0 +1,5 @@
+package com.github.panpf.zoomimage.compose.common.test.util
+
+class KeyHandlerTest {
+    // TODO test
+}