+++ zoomimage-compose/src/commonTest/kotlin/com/github/panpf/zoomimage/compose/common/test/zoom/KeyZoomTest.kt
@@ -1,0 +1,5 @@
+package com.github.panpf.zoomimage.compose.common.test.zoom
+
+class KeyZoomTest {
+    // TODO test
+}